sink(ticdc): split interface.go
